requirements.txt
Flask==3.1.1
python-dotenv==1.0.0
flask-cors==4.0.0
supabase==2.3.4
